rename dataquery.gen.ts as dataquery.ts
